Fix Recently Active timestamps with shared relative-time formatter
