+++ agent-command-center/src/utils/formatLastActive.ts
@@ -1,0 +1,14 @@
+export const formatLastActive = (date?: Date): string => {
+  if (!date) return 'Never';
+
+  const now = new Date();
+  const diff = now.getTime() - date.getTime();
+  const minutes = Math.floor(diff / 60000);
+  const hours = Math.floor(minutes / 60);
+  const days = Math.floor(hours / 24);
+
+  if (minutes < 1) return 'Just now';
+  if (minutes < 60) return `${minutes}m ago`;
+  if (hours < 24) return `${hours}h ago`;
+  return `${days}d ago`;
+};
